Move session markup to a separate file
